move vpn banner server code to own module
