+++ template-helpers/header.js
@@ -2,8 +2,6 @@
 
 const { getStrings, getFxaUrl } = require("./hbs-helpers");
 const { LocaleUtils } = require("./../locale-utils");
-const fs = require("fs");
-const Reader = require("@maxmind/geoip2-node").Reader;
 
 function getSignedInAs(args) {
   const locales = args.data.root.req.supportedLocales;
@@ -64,29 +62,8 @@ function fxaMenuLinks(args) {
   return getStrings(fxaLinks, locales);
 }
 
-function getIpInfo(args) {
-  const dbBuffer = fs.readFileSync("./tests/mmdb/GeoLite2-City-Test.mmdb");
-  const reader = Reader.openBuffer(dbBuffer);
-  const clientIp = args.data.root.constants.NODE_ENV === "dev" ? "2.125.160.216" : args.data.root.req.ip; // TODO: normalize IP for different ip4/ip6 formats
-  const info = {ip: clientIp};
-  let geoData, city, stateOrCountry;
-
-  try{
-    geoData = reader.city(clientIp);
-    city = geoData.city.names.en || ""; // TODO: add optional chaining after Node version upgrade
-    stateOrCountry = geoData.subdivisions[0].isoCode || geoData.country.isoCode || ""; // TODO: add optional chaining after Node version upgrade
-  }catch(e){
-    console.warn(e);
-  }
-
-  info.location = `${city}${stateOrCountry ? `, ${stateOrCountry}` : ""}`;
-
-  return info;
-}
-
 module.exports = {
   navLinks,
   fxaMenuLinks,
   getSignedInAs,
-  getIpInfo,
 };
