+++ template-helpers/index.js
@@ -12,6 +12,7 @@ const homepage = require("./homepage");
 const legacyHelpers = require("./hbs-helpers");
 const scanResults = require("./scan-results");
 const signUpBanners = require("./sign-up-banners");
+const vpnBanner = require("./vpn-banner");
 
 module.exports = {
   helpers: Object.assign(
@@ -26,6 +27,7 @@ module.exports = {
     homepage,
     legacyHelpers,
     scanResults,
-    signUpBanners
+    signUpBanners,
+    vpnBanner
   ),
 };
